Big AdminDialog update.
Updated Admin Dialog to be able to almost delete and add users just need
the querys for it, Also made it so that admins could create or delete
rentals for it, also need querys for it. Updated database so that it has
a returned date so we can query for non returned movies! WEOW! wut an
updates