Main Edits to AdminDialog & RentalDialog
AdminDialog:
-Deleted unneeded components.
- updated sizes so as to see you better my dear...(minimum window size
updated for better fitting)
- changed default button to "rentals in last 24 hours" so as to make
buttons be off by default.
- Admins can now delete rentals from the rental list, they can also now
delete users ( only users) from the database.
-Layed the framework for returning movies, we should maybe make a
trigger (i think in progress?) so that when a rental is returned we will
update the assigned users quota and update the amount avalible in the in
the rentals table.

-changed layout for viewing pleasure.

CreateRentalDialog:
- Added a date spinner to allow setting of release dates.
- Allow release dates to be set in the database now, instead of just
passing the current date
- Updated bad completion dialog messages to show correct text now.
-changed layout a little for viewing pleasure, buttons are now on the
bottom

Create User Dialog:
-changed layout a little for viewing pleasure, buttons are now on the
bottom

Main Gui:
- updated sizes so as to see you better my dear...(minimum window size
updated)